Refactor into an SMB server relay mixin
