+++ lib/msf/core/exploit/remote/smb/relay/ntlm/server_client.rb
@@ -25,7 +25,7 @@ module Msf::Exploit::Remote::SMB::Relay::NTLM
       session.metadata[:relay_target] = @relay_targets.next(session.metadata[:identity])
       # If there's no more targets to relay to, just tree connect to the currently running server instead
       if session.metadata[:relay_target].nil?
-        logger.print_status("identity: #{session.metadata[:identity]} - All targets relayed to")
+        logger.print_status("Identity: #{session.metadata[:identity]} - All targets relayed to")
         return super(request, session)
       end
 
@@ -162,7 +162,7 @@ module Msf::Exploit::Remote::SMB::Relay::NTLM
         @relay_targets.on_relay_end(relayed_connection.target, identity: session.metadata[:identity], is_success: is_success)
 
         if is_success
-          logger.print_good("identity: #{session.metadata[:identity]} - Successfully authenticated against relay target #{display_target(relayed_connection.target)}")
+          logger.print_good("Identity: #{session.metadata[:identity]} - Successfully authenticated against relay target #{display_target(relayed_connection.target)}")
           session.metadata[:incoming_challenge_response] = ntlm_message
 
           @listener.on_ntlm_type3(
@@ -171,19 +171,19 @@ module Msf::Exploit::Remote::SMB::Relay::NTLM
             ntlm_type2: session.metadata[:relay_target_server_challenge],
             ntlm_type3: session.metadata[:incoming_challenge_response]
           )
-          @listener.on_relay_success(relay_connection: relayed_connection)
+          @listener.on_relay_success(relay_connection: relayed_connection, relay_identity: session.metadata[:identity])
         else
           @listener.on_relay_failure(relay_connection: relayed_connection)
           relayed_connection.disconnect!
 
           if relay_result.nt_status == WindowsError::NTStatus::STATUS_LOGON_FAILURE
-            logger.print_warning("identity: #{session.metadata[:identity]} - Relay failed due to client authentication details not matching any account on target server #{display_target(relayed_connection.target)}")
+            logger.print_warning("Identity: #{session.metadata[:identity]} - Relay failed due to client authentication details not matching any account on target server #{display_target(relayed_connection.target)}")
           else
             error_code = WindowsError::NTStatus.find_by_retval(relay_result.nt_status.value).first
             if error_code.nil?
-              logger.print_warning("identity: #{session.metadata[:identity]} - Relay against target #{display_target(relayed_connection.target)} failed with unexpected error: #{relay_result.nt_status.value}")
+              logger.print_warning("Identity: #{session.metadata[:identity]} - Relay against target #{display_target(relayed_connection.target)} failed with unexpected error: #{relay_result.nt_status.value}")
             else
-              logger.print_warning("identity: #{session.metadata[:identity]} - Relay against target #{display_target(relayed_connection.target)} failed with unexpected error: #{error_code.name}: #{error_code.description}")
+              logger.print_warning("Identity: #{session.metadata[:identity]} - Relay against target #{display_target(relayed_connection.target)} failed with unexpected error: #{error_code.name}: #{error_code.description}")
             end
           end
